refactor(stack): move responsive properties
