+++ packages/admin-ui/src/stack/index.ts
@@ -1,1 +1,1 @@
-export { Stack, useStack, StackProps } from './stack'
+export * from './stack'
